chore: openjdk pkgs
fix openjdk specs

+++ oreon-11-rp1/java-25-openjdk-portable/java-25-openjdk-portable.spec
@@ -1205,12 +1205,18 @@ if [ ! -x "${rev_boot}/bin/java" ] && [ -x /usr/lib/jvm/java-%{buildjdkver}-open
   rev_boot=/usr/lib/jvm/java-%{buildjdkver}-openjdk-fastdebug
 fi
 if [ -x "${rev_boot}/bin/java" ]; then
-  mkdir -p .src-rev-build
-  pushd .src-rev-build
-  sh ../%{top_level_dir_name}/configure --with-boot-jdk="${rev_boot}" --with-boot-jdk-jvmargs="-XX:+UseParallelGC"
-  make store-source-revision
-  popd
-  rm -rf .src-rev-build
+  if [ -f %{top_level_dir_name}/.src-rev ]; then
+    :
+  elif [ -d %{top_level_dir_name}/.git ]; then
+    mkdir -p .src-rev-build
+    pushd .src-rev-build
+    sh ../%{top_level_dir_name}/configure --with-boot-jdk="${rev_boot}" --with-boot-jdk-jvmargs="-XX:+UseParallelGC"
+    make store-source-revision
+    popd
+    rm -rf .src-rev-build
+  elif [ -n "%{?source_revision}" ]; then
+    printf '.:git:%{source_revision}\n' > %{top_level_dir_name}/.src-rev
+  fi
 fi
 
 %build
